pyproject.toml
[project]
name = "northern-lights"
version = "0.1.0"
description = "Aurora Borealis visibility tracker with email notifications"
authors = [
    {name = "Edgar Gonçalves", email = "edgar.g.goncalves@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.3",
    "geopy>=2.4.1",
    "rich>=13.9.4",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utils"]
